moved compressFile into separate module to reference from worker
